requirements.txt
# Dev environment
pip
autopep8

# App
flask
Werkzeug
python-dateutil